+++ green-20220302/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-07-28 Version: 3.5.2
+- Update API DescribeFileModerationResult: add response parameters Body.Data.PageSummary.RiskSummary.
+- Update API MultiModalGuardForBase64: add request parameters FileBase64Str.
+
+
 2026-07-16 Version: 3.5.1
 - Update API MultiModalGuardAsyncResult: add response parameters Body.Data.FrameResult.Frames.$.ErrorCode.
 - Update API MultiModalGuardAsyncResult: add response parameters Body.Data.FrameResult.Frames.$.ErrorMsg.